Remove useless component
